+++ src/models/his/his_hosxpv4.ts
@@ -1567,6 +1567,13 @@ export class HisHosxpv4Model {
     }
 
     // MOPH ERP
+    countBedNo(db: Knex) {
+        return db('bedno').count('bedno.bedno as total_bed')
+            .leftJoin('roomno', 'bedno.roomno', 'roomno.roomno')
+            .leftJoin('ward', 'roomno.ward', 'ward.ward')
+            .where('ward.ward_active', 'Y').first();
+    }
+
     getBedNo(db: Knex, bedno: any = null) {
         let sql = db('bedno')
             .leftJoin('roomno', 'bedno.roomno', 'roomno.roomno')