+++ src/models/his/his_mitnet.ts
@@ -3,7 +3,7 @@ const maxLimit = 250;
 const hcode = process.env.HOSPCODE;
 let hisHospcode = process.env.HOSPCODE;
 
-export class HisModel {
+export class HisMitnetModel {
   check() {
     return true;
   }